+++ lib/src/network/enum_post_type.dart
@@ -1,4 +1,0 @@
-/// (en) Specifies the encode type for simple POST.
-///
-/// (ja) 簡易なPOSTの際のエンコードタイプを指定します。
-enum EnumPostEncodeType { urlEncoded, json }