+++ lib/src/static_fields/f_json_keys_from_server.dart
@@ -1,9 +1,9 @@
 /// サーバーからのデータ受信時にJSONで利用されるキーを定義したクラス。
 class FJsonKeysFromServer {
-  static String accessToken = "access_token";
-  static String tokenType = "token_type";
-  static String expiresIn = "expires_in";
-  static String refreshToken = "refresh_token";
+  static const String accessToken = "access_token";
+  static const String tokenType = "token_type";
+  static const String expiresIn = "expires_in";
+  static const String refreshToken = "refresh_token";
   // 要求したスコープと差異がある場合などに必要になる。
-  static String scope = "scope";
+  static const String scope = "scope";
 }