+++ lib/src/tools/error_reporter_web.dart
@@ -1,2 +1,0 @@
-export 'error_report_obj.dart';
-export 'error_reporter.dart';